Convert Hounsfield values from integer to shorts.

+++ ReConstruct.Data.Dicom/dicom.fs
@@ -49,7 +49,7 @@ type SliceLayout =
 type ImageSlice =
     {
         Layout: SliceLayout;
-        HField: int[];
+        HField: int16[];
     }
 
 type DicomTree =
