+++ ReConstruct.UI.Views/app-state.fs
@@ -3,4 +3,4 @@
 open ReConstruct.Data.Dicom
 
 module AppState =
-    let mutable Level: Option<int> = Some Imaging.BONES_ISOVALUE
+    let mutable Level: Option<int16> = Some Imaging.BONES_ISOVALUE
